Update PacketHandlerModern.java
Adding new 1.20 - 1.20.1 field because of Minecraft obfuscation 
Old 1.19.x-1.19.4 uses field "ac" for field to determine whether the server is actually being setup but in 1.20.x it's different field beacuse of the obfuscation. 

And because there wasn't any `case 20:`, the while loop never actually ended thus never being able to reach the Packet Handler setup part. 

This change fixes that.